refactor: move data classes to dedicated directory and split up jsonconf classes
