+++ ulauncher/data/__init__.py
@@ -1,0 +1,5 @@
+from ulauncher.data.base_data_class import BaseDataClass
+from ulauncher.data.json_conf import JsonConf
+from ulauncher.data.json_key_value_conf import JsonKeyValueConf
+
+__all__ = ["BaseDataClass", "JsonConf", "JsonKeyValueConf"]
