+++ ulauncher/data/base_data_class.py
