+++ ulauncher/data/json_key_value_conf.py
@@ -4,75 +4,14 @@ import logging
 from pathlib import Path
 
 # MutableMapping from typing (not collections.abc) supports subscript syntax on Python 3.8 for class bases.
-# MutableMappingABC from collections.abc is used for isinstance checks.
 from typing import Any, Callable, Generic, Iterator, MutableMapping, TypeVar, cast, get_args, get_origin
 
-from ulauncher.utils.base_data_class import BaseDataClass
-from ulauncher.utils.json_utils import json_load, json_save
+from ulauncher.data._file_cache import _load_cached_file_instance, _save_cached_file_instance
 
 logger = logging.getLogger()
-T = TypeVar("T", bound="JsonConf")
 V = TypeVar("V")
 K = TypeVar("K", bound=str)
 KVC = TypeVar("KVC", bound="JsonKeyValueConf[Any, Any]")
-FileInstanceT = TypeVar("FileInstanceT")
-_file_instances: dict[tuple[Path, type[object]], object] = {}
-_instance_paths: dict[int, Path] = {}
-
-
-def _load_cached_file_instance(cls: type[FileInstanceT], path: str | Path) -> tuple[FileInstanceT, Any]:
-    file_path = Path(path).resolve()
-    key = (file_path, cls)
-    data = json_load(file_path)
-    instance = cast("FileInstanceT", _file_instances.get(key))
-    if instance is None:
-        instance = cls()
-        _file_instances[key] = instance
-        _instance_paths[id(instance)] = file_path
-    return instance, data
-
-
-def _save_cached_file_instance(instance: object, data: Any, *, sort_keys: bool = True) -> bool:
-    file_path = _instance_paths.get(id(instance))
-    if file_path is None:
-        logger.error("Could not resolve file path for instance %s", instance.__class__.__name__)
-        return False
-
-    return json_save(data, file_path, sort_keys=sort_keys)
-
-
-class JsonConf(BaseDataClass):
-    """
-    JsonConf
-
-    This is an helper class to handle json config files
-    It has two convenient methods for loading and saving the data, which also depuplicate class instances,
-    So that if you create two instances for the same file it will actually reuse the first instance
-    (to avoid overwriting state changes)
-
-    File paths are stored in an external reference object, which is not part of the data.
-    """
-
-    def __setitem__(self, key: str, value: dict[str, Any], validate_type: bool = True) -> None:
-        if hasattr(self.__class__, key):
-            class_val = getattr(self.__class__, key)
-            # WARNING: this logic does not work with union types or nullable types,
-            # since it derives the type from the initial class value
-            if validate_type and not isinstance(value, type(class_val)):
-                msg = f'"{key}" must be of type {type(class_val).__name__}, {type(value).__name__} given.'
-                raise KeyError(msg)
-
-        super().__setitem__(key, value)
-
-    @classmethod
-    def load(cls: type[T], path: str | Path) -> T:
-        instance, data = _load_cached_file_instance(cls, path)
-        instance.update(data)
-        return instance
-
-    def save(self, *args: Any, **kwargs: Any) -> bool:
-        self.update(*args, **kwargs)
-        return _save_cached_file_instance(self, self)
 
 
 class JsonKeyValueConf(MutableMapping[str, V], Generic[K, V]):
